Move LayerPreview into renderer
