+++ packages/noya-renderer/src/index.ts
@@ -7,6 +7,7 @@ export { default as SketchFileRenderer } from './components/SketchFileRenderer';
 export { default as SketchLayer } from './components/layers/SketchLayer';
 export { default as SketchArtboard } from './components/layers/SketchArtboard';
 export { default as SketchGroup } from './components/layers/SketchGroup';
+export { default as LayerPreview } from './components/LayerPreview';
 
 export type { Context };
 export { uuid, Primitives };
